Refactor: Fix type conversions and signedness warnings across libft and parser

+++ libft/convert/ft_itoa.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: flink <flink@student.42.fr>                +#+  +:+       +#+        */
+/*   By: flenski <flenski@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/30 10:39:29 by flenski           #+#    #+#             */
-/*   Updated: 2026/02/18 10:43:26 by flink            ###   ########.fr       */
+/*   Updated: 2026/06/17 12:37:35 by flenski          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@ char	*ft_itoa(long long n)
 	if (n < 0)
 		nb = -(unsigned long long)n;
 	else
-		nb = n;
-	str = malloc(ft_numlen(n) + (n < 0) + 1);
+		nb = (unsigned long long)n;
+	str = malloc(ft_numlen(n) + (size_t)(n < 0) + 1);
 	if (!str)
 		return (NULL);
 	if (nb == 0)
